requirements.txt
# TeleForward - Telegram to Discord Forwarder
# Python 3.10+ required

# Telegram API
telethon>=1.34.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# HTTP client for Discord webhooks
httpx>=0.27.0

# Environment variable management
python-dotenv>=1.0.0

# Async utilities
anyio>=4.0.0

# TUI
rich>=13.7.0
